# apps/api/src/routes/v1/openapi.yaml
openapi: 3.0.3
info:
  title: OpenMeet Public API
  version: 1.0.0
  description: |
    Enterprise REST API for meeting intelligence platform.

    ## Authentication
    All endpoints require an API key passed via the `Authorization` header:
    ```
    Authorization: Bearer nb_your_api_key_here
    ```

    ## Rate Limiting
    Rate limits are tiered based on your subscription:
    - **Free**: 100 requests/hour
    - **Pro**: 1,000 requests/hour
    - **Business**: 5,000 requests/hour
    - **Enterprise**: 10,000 requests/hour

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Your rate limit
    - `X-RateLimit-Remaining`: Remaining requests in window
    - `X-RateLimit-Reset`: When the limit resets (ISO 8601)

    ## Webhooks
    Subscribe to events via webhooks with HMAC-SHA256 signature verification.

  contact:
    name: API Support
    email: api@openmeet.com
    url: https://docs.openmeet.com
  license:
    name: Proprietary
    url: https://openmeet.com/terms
servers:
  - url: https://api.openmeet.com/api/v1
    description: Production
  - url: https://staging-api.openmeet.com/api/v1
    description: Staging
security:
  - BearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      description: Get paginated list of users in organization
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: role
          in: query
          schema:
            type: string
            enum: [user, admin, super_admin]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user
      description: Retrieve detailed user information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'

    patch:
      tags: [Users]
      summary: Update user
      description: Update user information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key with 'ff_' prefix
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin]
        createdAt:
          type: string
          format: date-time
    UsersListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
    UserDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        avatar:
          type: string
          format: uri
        timezone:
          type: string
        language:
          type: string
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
        nextOffset:
          type: integer
          nullable: true
...